backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1
rq==1.16.2

# Auth
pyjwt==2.10.1
argon2-cffi==23.1.0
python-multipart==0.0.18

# Validation & serialization
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# HTTP client (for external soccer API)
httpx==0.28.1

# Image processing (avatar validation)
Pillow==11.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Linting
ruff==0.8.6
